+++ examples/basic.rs
@@ -13,8 +13,8 @@ struct Name(String);
 struct Depth(u32);
 
 pub struct SizeSystem;
-impl System<EntityStore, PhantomContext> for SizeSystem {
-    fn run(&self, ecm: &mut EntityComponentManager<EntityStore>) {
+impl System<EntityStore> for SizeSystem {
+    fn run(&self, ecm: &mut EntityComponentManager<EntityStore>, _: &mut Resources) {
         let (e_store, c_store) = ecm.stores_mut();
 
         for entity in &e_store.inner {
@@ -27,8 +27,8 @@ impl System<EntityStore, PhantomContext> for SizeSystem {
 }
 
 pub struct PrintSystem;
-impl System<EntityStore, PhantomContext> for PrintSystem {
-    fn run(&self, ecm: &mut EntityComponentManager<EntityStore>) {
+impl System<EntityStore> for PrintSystem {
+    fn run(&self, ecm: &mut EntityComponentManager<EntityStore>, _: &mut Resources) {
         let (e_store, c_store) = ecm.stores_mut();
 
         for entity in &e_store.inner {