+++ examples/init_cleanup.rs
@@ -2,24 +2,24 @@ use dces::prelude::*;
 
 struct InitSystem;
 
-impl System<EntityStore, PhantomContext> for InitSystem {
-    fn run(&self, _: &mut EntityComponentManager<EntityStore>) {
+impl System<EntityStore> for InitSystem {
+    fn run(&self, _: &mut EntityComponentManager<EntityStore>, _: &mut Resources) {
         println!("Init");
     }
 }
 
 struct CleanupSystem;
 
-impl System<EntityStore, PhantomContext> for CleanupSystem {
-    fn run(&self, _: &mut EntityComponentManager<EntityStore>) {
+impl System<EntityStore> for CleanupSystem {
+    fn run(&self, _: &mut EntityComponentManager<EntityStore>, _: &mut Resources) {
         println!("Cleanup");
     }
 }
 
 struct PrintSystem;
 
-impl System<EntityStore, PhantomContext> for PrintSystem {
-    fn run(&self, _: &mut EntityComponentManager<EntityStore>) {
+impl System<EntityStore> for PrintSystem {
+    fn run(&self, _: &mut EntityComponentManager<EntityStore>, _: &mut Resources) {
         println!("Print");
     }
 }