+++ src/lib.rs
@@ -27,8 +27,8 @@
 //!
 //! struct PrintSystem;
 //!
-//! impl System<EntityStore, PhantomContext> for PrintSystem {
-//!    fn run(&self, ecm: &mut EntityComponentManager<EntityStore>) {
+//! impl System<EntityStore> for PrintSystem {
+//!    fn run(&self, ecm: &mut EntityComponentManager<EntityStore>, _res: &mut Resources) {
 //!        let (e_store, c_store) = ecm.stores();
 //!
 //!        for entity in &e_store.inner {
@@ -62,5 +62,6 @@ pub mod component;
 pub mod entity;
 pub mod error;
 pub mod prelude;
+pub mod resources;
 pub mod system;
 pub mod world;